Updates to main THC function.
Working great on my pllasma cutter. I noticed I was using delay ms and not us(microseconds) for the direction and step delays. It works much better now.

+++ Grbl_Esp32/i2s_out.h
@@ -167,7 +167,7 @@ void i2s_out_delay();
    Set the pulse callback period in microseconds
    (like the timer period for the ISR)
  */
-int i2s_out_set_pulse_period(uint32_t period);
+int i2s_out_set_pulse_period(uint64_t period);
 
 /*
    Register a callback function to generate pulse data
